sms repo

+++ Hot4/Hot4.ViewModel/SMSSearchModel.cs
@@ -7,7 +7,7 @@
         public string Mobile { get; set; }
         public string MessageText { get; set; }
         public int SmppId { get; set; }
-        public byte StateId { get; set; }
+        public int StateId { get; set; }
         public int PageNo { get; set; }
         public int PageSize { get; set; }
     }
